+++ src/Base/AbstractResponse.php
@@ -180,32 +180,6 @@ abstract class AbstractResponse implements ResponseInterface
     }
 
     /**
-     * Sets the expiry time.
-     *
-     * @since      1.0.0
-     * @deprecated Use setExpiryDateTime instead.
-     *
-     * @param DateTimeImmutable|null $expiry The expiry time or null for immediate expiry.
-     */
-    public function setExpiry(?DateTimeImmutable $expiry = null): void
-    {
-        $date = new DateTimeImmutable();
-        $expiry = $expiry ?: $date;
-
-        $this->setHeader('Date', $date->setTimezone(new DateTimeZone('UTC'))->format('D, d M Y H:i:s \G\M\T'));
-        $this->setHeader('Expires', $expiry->setTimezone(new DateTimeZone('UTC'))->format('D, d M Y H:i:s \G\M\T'));
-
-        if ($expiry <= $date) {
-            $this->setHeader('Cache-Control', 'no-cache, no-store, must-revalidate, max-age=0');
-
-            return;
-        }
-
-        $maxAge = $expiry->getTimestamp() - $date->getTimestamp();
-        $this->setHeader('Cache-Control', 'public, max-age=' . $maxAge);
-    }
-
-    /**
      * Sets the expiry date time.
      *
      * @since 2.2.0